♻️  refactor: S3 관련 클래스 external 모듈로 분리 #124
