+++ motivoo-api/src/test/java/controller/BaseControllerTest.java
@@ -29,8 +29,8 @@ import sopt.org.motivoo.domain.auth.service.apple.AppleClaimsValidator;
 import sopt.org.motivoo.domain.auth.service.apple.AppleLoginService;
 import sopt.org.motivoo.domain.external.firebase.FirebaseService;
 import sopt.org.motivoo.domain.external.firebase.config.FirebaseConfig;
-import sopt.org.motivoo.domain.external.s3.S3Service;
-import sopt.org.motivoo.domain.external.s3.config.AWSConfig;
+import sopt.org.motivoo.external.s3.S3Service;
+import sopt.org.motivoo.external.s3.config.AWSConfig;
 import sopt.org.motivoo.domain.external.slack.SlackService;
 
 @AutoConfigureMockMvc
